Rename FullDependencyRule -> DependencyRule
