+++ src/types/config/DependencyRule.ts
@@ -1,4 +1,4 @@
-export default interface FullDependencyRule {
+export default interface DependencyRule {
     dependency: string;
     accessibleTo: string | string[];
 };
