+++ src/types/rawConfig/RawDependencyRule.ts
@@ -1,4 +1,4 @@
-import FullDependencyRule from '../FullDependencyRule';
+import FullDependencyRule from '../config/DependencyRule';
 
 type RawDependencyRule = FullDependencyRule | string;
 export default RawDependencyRule;
